+++ libs/common/src/lib/interfaces/responses/portfolio-performance-response.interface.ts
@@ -4,6 +4,12 @@ import { ResponseError } from './errors.interface';
 
 export interface PortfolioPerformanceResponse extends ResponseError {
   chart?: HistoricalDataItem[];
+  dateOfFirstActivity: Date;
+
+  /**
+   * @deprecated Use `dateOfFirstActivity` instead
+   */
   firstOrderDate: Date;
+
   performance: PortfolioPerformance;
 }